requirements.txt
numpy
tensorflow
scipy
pillow
pandas
scikit-learn
scikit-fuzzy
torch
torchvision
